web_login/aliexpress/requirements.txt
PyExecJS==1.5.1
requests==2.19.1
